Fix compilation errors: ApiResponse ambiguous import and AchievementService method calls

+++ src/main/java/com/swp391project/SWP391_QuitSmoking_BE/dto/health/HealthMetricDTO.java
@@ -24,7 +24,7 @@ public class HealthMetricDTO {
     private LocalDateTime targetDate;
     private LocalDateTime achievedDate;
     private Boolean isCompleted;
-    private Long timeRemainingHours;
+    private Double timeRemainingHours;
     private String timeRemainingFormatted;
     private Boolean hasRegressed;
     private LocalDateTime createdAt;
